Extract Left Nav component; refactor folders
