+++ src/components/vars.sass
@@ -1,2 +1,4 @@
 $border-radius: 20px
 $border-width: 1px
+$class-work-title-bar-height: 34px
+$class-work-left-nav-width: 180px
