+++ src/components/view-class-work/view-class.sass
@@ -1,7 +1,4 @@
-@import "vars"
-
-$titleBarHeight: 34px
-$leftNavWidth: 180px
+@import "../vars"
 
 .viewClass
   position: fixed
@@ -19,14 +16,14 @@ $leftNavWidth: 180px
     top: 0
     right: 0
     left: 0
-    height: $titleBarHeight
+    height: $class-work-title-bar-height
     background-color: #9fc5e8
     border-bottom: 1px solid #000
     border-radius: $border-radius $border-radius 0 0
     user-select: none
 
     .titleBarContents
-      height: $titleBarHeight
+      height: $class-work-title-bar-height
       font-size: 1.3em
       display: flex
       justify-content: space-around
@@ -51,37 +48,11 @@ $leftNavWidth: 180px
         width: 1.3em
         margin: 0 10px 0 5px
 
-  .leftNav
-    position: absolute
-    top: $titleBarHeight + $border-width
-    left: 0
-    bottom: 0
-    width: $leftNavWidth
-    background-color: #dae6f3
-    border-right: 1px solid #000
-    border-radius: 0 0 0 $border-radius
-    user-select: none
-    overflow: auto
-
-    .leftNavContents
-      padding: 20px
-
-      ul
-        margin: 0
-        padding: 0
-        list-style: none
-        font-size: 1.1em
-
-        .selected
-          font-weight: bold
-        li:hover
-          text-decoration: underline
-
   .viewer
     position: absolute
-    top: $titleBarHeight + $border-width
+    top: $class-work-title-bar-height + $border-width
     bottom: 0
-    left: $leftNavWidth + $border-width
+    left: $class-work-left-nav-width + $border-width
     right: 0
 
     .viewerMessage
